+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Check [README.md](README.md#installation-on-helm-2)
 - Implement alternate syntax (https://github.com/jkroepke/helm-secrets/pull/52)
 - Remote values support (supporting http:// and helm downloader plugins) (https://github.com/jkroepke/helm-secrets/pull/54)
 - Let downloader plugin supports remote files and all secrets drivers (https://github.com/jkroepke/helm-secrets/pull/55)
+- Externalize custom vault driver logic. (https://github.com/jkroepke/helm-secrets/pull/63)
 
 ### Fixes
 - Vault driver: If vault command failed, the script execution was not terminated. (https://github.com/jkroepke/helm-secrets/pull/61)